refactor(runs): extract observation bundle serialization into core service
The runs export/import command owned the entire portable-bundle format:
building a bundle from the store, writing it to a directory, reading and
validating it back, artifact-byte packaging, and directory-artifact
zip/extract. All of that is direct filesystem mutation + file-writer
construction living in the command layer, tripping the
thin_command_adapter audit detector.

Move the bundle format into a new core::observation::bundle service: the
bundle types (manifest, bundle, artifact-bytes), build_bundle, the
portable-artifact packaging, write/read_bundle_dir, and the zip/extract
+ json IO helpers. The command keeps CLI args, run selection, and record
remapping, delegating every filesystem operation to core.

Command module drops from 966 to ~330 lines; the bundle format is now a
reusable core primitive.

Clears the runs/bundle.rs thin_command_adapter_violation.
